Organize management work into separate modules
This collects server state into an InstanceManager object, and
migrates many of the other resources to other modules.
