+++ AdminServer/appscale/admin/instance_manager/__init__.py
@@ -1,0 +1,1 @@
+from appscale.admin.instance_manager.instance_manager import InstanceManager
