+++ AdminServer/appscale/admin/instance_manager/utils.py
@@ -7,10 +7,10 @@ import os
 import shutil
 import subprocess
 
+from appscale.admin.constants import InvalidSource
+from appscale.admin.instance_manager.constants import (
+  CONFLICTING_JARS, LOGROTATE_CONFIG_DIR, MODIFIED_JARS, MONIT_INSTANCE_PREFIX)
 from appscale.common.constants import CONFIG_DIR
-from .constants import CONFLICTING_JARS
-from .constants import MODIFIED_JARS
-from ..constants import InvalidSource
 
 logger = logging.getLogger('appscale-admin')
 
@@ -101,3 +101,61 @@ def remove_conflicting_jars(source_path):
     for pattern in CONFLICTING_JARS:
       if fnmatch.fnmatch(file, pattern):
         os.remove(os.path.join(lib_dir, file))
+
+
+def remove_logrotate(app_name):
+  """ Removes logrotate script for the given application.
+
+  Args:
+    app_name: A string, the name of the application to remove logrotate for.
+  Returns:
+    True on success, False otherwise.
+  """
+  app_logrotate_script = "{0}/appscale-{1}".\
+    format(LOGROTATE_CONFIG_DIR, app_name)
+  logging.debug("Removing script: {}".format(app_logrotate_script))
+
+  try:
+    os.remove(app_logrotate_script)
+  except OSError:
+    logging.error("Error deleting {0}".format(app_logrotate_script))
+    return False
+
+  return True
+
+
+def setup_logrotate(app_name, log_size):
+  """ Creates a logrotate script for the logs that the given application
+      will create.
+
+  Args:
+    app_name: A string, the application ID.
+    log_size: An integer, the size of logs that are kept per application server.
+      The size should be in bytes.
+  Returns:
+    True on success, False otherwise.
+  """
+  # Write application specific logrotation script.
+  app_logrotate_script = "{0}/appscale-{1}".\
+    format(LOGROTATE_CONFIG_DIR, app_name)
+
+  log_prefix = ''.join([MONIT_INSTANCE_PREFIX, app_name])
+
+  # Application logrotate script content.
+  contents = """/var/log/appscale/{log_prefix}*.log {{
+  size {size}
+  missingok
+  rotate 7
+  compress
+  delaycompress
+  notifempty
+  copytruncate
+}}
+""".format(log_prefix=log_prefix, size=log_size)
+  logging.debug("Logrotate file: {} - Contents:\n{}".
+    format(app_logrotate_script, contents))
+
+  with open(app_logrotate_script, 'w') as app_logrotate_fd:
+    app_logrotate_fd.write(contents)
+
+  return True
