+++ common/appscale/common/constants.py
@@ -6,6 +6,7 @@ from kazoo.retry import KazooRetry
 
 
 class HTTPCodes(object):
+  OK = 200
   BAD_REQUEST = 400
   UNAUTHORIZED = 401
   FORBIDDEN = 403
